[hayden] feat(added new metrics and significance_test function)
New metrics: Hellinger Distance, Bhattacharyya Distance, Chi-Squared Statistic

New function: Added function to test significance of an average of distance metrics (averaged across multiple runs of the QGB to account for stochastic uncertainty), to conclude if the distance is statistically significant.
This function includes the option to perform a chi-squared test with chi-squared statistics.

Note: some unit tests done but many more to be performed, significance_test() is untested. Consider view-only for now.